changed function name
Changed functions to async
