--- app/player-stuff.js
@@ -1,6 +1,0 @@
-async function getPlayer(req, res) {
-    const profile = { id: req.params.id }; //some data base stuff here
-    res.status(200).send(profile);
-}
-
-module.exports = { getPlayer };
