+++ app/sql_search.js
@@ -1,0 +1,40 @@
+const mysql = require('mysql');
+
+var con = mysql.createConnection({
+	host: 'localhost',
+	user: 'standarduser',
+	password: "",
+	database: 'harvold'
+});
+
+con.connect(function(err)
+{
+	if (err) 
+		throw err;
+	console.log("MySQL Connected");
+});
+
+async function getPlayer(req, res) {
+	var user = req.params.username;
+	
+	var sql = "SELECT first_name, last_name, username, status FROM users WHERE username= ?";
+	con.query(sql, [user], function(err, result)
+	{
+		if (err) throw err;
+		console.log("player_search");
+		res.status(200).json(result);
+	});
+    
+}
+
+async function getPokemon(req, res){
+	var user = req.query['username'];
+	var sql = "SELECT name, hp, max_hp, exp, to_next FROM pokemon WHERE owner= ?";
+	con.query(sql, [user], function(err, result)
+	{
+		if (err) throw err;
+		console.log ("searched");
+		res.status(200).json(result);
+	});
+}
+module.exports = { getPlayer };
